--- src/components/Form/Form.module.css
@@ -1,7 +1,0 @@
-.form {
-  background: #ffffff;
-  border-radius: 3px;
-  max-width: 450px;
-  width: 100%;
-  padding: 1em 2em;
-}
