+++ src/components/Login/Signin/Signin.module.css
@@ -22,13 +22,13 @@
   font-weight: 400;
   border-radius: 5px;
   color: #fff;
-  background-color: #0171d3;
+  background-color: rgb(255, 114, 114);
   transition: all 0.3s ease;
   cursor: pointer;
 }
 
 .field button:hover {
-  background-color: #016dcb;
+  background-color: rgba(255, 114, 114, 0.9);
 }
 
 .formLink {
