+++ src/components/Login/Signin/index.jsx
@@ -1,17 +1,16 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import Input from "../Input";
-import ToggleSign from "../ToggleSign";
-import styles from "./Signup.module.css";
 
-export default function Signup({ toggleSign }) {
+import ToggleSign from "../ToggleSign";
+import styles from "./Signin.module.css";
+
+export default function Signin({ isLogin, toggleSign }) {
   return (
     <>
       <form className={styles.form}>
         <Input type="text" placeholder="ID" />
-        <Input type="email" placeholder="Email" />
         <Input type="password" placeholder="Password" />
-        <Input type="password" placeholder="Confirm Password" />
         <div className={styles.formLink}>
           <Link className={styles.forgotPW}>아이디 찾기</Link>
           <Link className={styles.forgotPW}>비밀번호 찾기</Link>
@@ -20,7 +19,7 @@ export default function Signup({ toggleSign }) {
           <button>Login</button>
         </div>
       </form>
-      <ToggleSign toggleSign={toggleSign} />
+      <ToggleSign isLogin={isLogin} toggleSign={toggleSign} />
     </>
   );
 }
