+++ src/components/Login/ToggleSign/index.jsx
@@ -1,12 +1,12 @@
 import React from "react";
 import styles from "./ToggleSign.module.css";
-export default function ToggleSign({ toggleSign }) {
+export default function ToggleSign({ isLogin, toggleSign }) {
   return (
     <div className={styles.formLink}>
       <span>
-        계정이 없습니까?{" "}
+        {isLogin ? "계정이 없습니까? " : "이미 회원입니까? "}
         <span onClick={toggleSign} className={styles.signupLink}>
-          Signup
+          {isLogin ? "회원가입" : "로그인"}
         </span>
       </span>
     </div>
