+++ src/index.js
@@ -1,17 +1,11 @@
-import { configureStore } from "@reduxjs/toolkit";
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { Provider } from "react-redux";
 import App from "./App";
 const root = ReactDOM.createRoot(document.getElementById("root"));
-const store = configureStore({
-  reducer: {},
-});
+
 root.render(
   <React.StrictMode>
-    <Provider store={store}>
-      <App />
-    </Provider>
+    <App />
   </React.StrictMode>
 );
 // If you want to start measuring performance in your app, pass a function
