+++ src/pages/Login/Login.module.css
@@ -26,6 +26,7 @@
   max-width: 450px;
   width: 100%;
   padding: 1em 2em;
+  box-shadow: rgba(0, 0, 0, 0.15) 1.95px 1.95px 2.6px;
 }
 
 .form .field {
